V3 (#53)
* Optimize Pads sample even more

* Rework pads module to be a class

* Add fast OBJ loader

* Change custom OBJ parser to Fast OBJ loader

* Add per-vertex color support using mtl

* Make bounding box a fixed array

* Implement support for multiple renderes and offline buffer allocation

* Renderer: Assign to white if no material

* Add multitexturing support (up to 10 textures per model)

* Add a new render sample

* Update modules descriptions on README

* Add font render class

* Fix compilation (#43)

* Emergency fix

* [makefile] migrate from bin2s to bin2c (#42)

* [makefile] migrate from bin2s to bin2c

* Update .gitignore

* Update embed.make

* Change some socket funcs

---------


* Use stat to get dir entry size

* Improve displayFunc functionality

* Add fast math functions

* Update compilation.yml

* Add vector basics

* Add second controller support

* Add more pad types support

* Add pad button events

* general path fixes

* feat: rework our 3D renderer to work under VU1 (#45)

* Add initial samples for VU1

* Add initial VU1 support

* Add batch logic to VU1 code

* Add workaround for waiting VU1

* Add printf again

* Add flush at the code start

* refactor: separate functions

* Change to use OpenVCL

* feat: add initial support for multiple mpgs

* Use VCL macros

* Add VU0 routines for lights

* change mpg name

* Add initial pipeline thing

* Do some general cleaning

* feat: diffuse and ambient lights

* perf: add normal math on VU1

* use dynamic batches

* feat: add no texture pipeline

* remove unused members

* feat: multitexturing on VU1

* clean temporary functions

* add linear filter support

* improve texture code readability

* improve bounding box calculation

* add multiple pipeline support

* add only color pipeline

* put obj in a class

* Add non-textured pipelines

* Update render.js

* change to not use image objects on render

* expose vertex access

* Add pipeline descriptors

* Add Render docs

* add screen docs

* add pads docs

* Update README.md

* Improve 3D camera

* optimize obj memory usage

* add optimize, size and bpp to image class

* add new image docs

* change render objects to hold js textures

* add get texture for render objects

* Update README.md

* optimize grayscale png loading

* expose vertex quantity

* Update README.md

* Add palette and pixels changing on Image

* add alpha support for models

* Update README.md

* fix file manager

* remove athena cli

* fix orbit camera

* add new camera functions

* change addvector to vu0

* add vector3 lib

* expose bounding boxes

* implement some 3d collision

* clean old unused code

* remove texture quantity limits for RenderObject

* remove fish model

* remove getFreeMemory

* remove leftover memory stuff

* save some kb with static functions

* remove vector leftovers

* Improve script post-run

* change some error types

* change athena error font

* dont preload fonts at start

* quantize dash icons

* optimize dash memory usage

* Add athena_restart for restart the sound without needed to reload, Add a workaround when sound end to just pause (to avoid stop working) (#48)

* Testing

* Try to never end wav reprodution (just pause)

* Implement restart on ath_sound, for restart a sound without reloading

* Fix build

* Fix build and rename function

* Implement athena_restart for ogg, fix arg count for athena_restart

* Introduze 'athena_get_position', 'athena_set_position', rename 'athena_duration' for 'athena_get_duration' (#49)

* Introduze 'athena_get_position', 'athena_set_position', rename 'athena_duration' for 'athena_get_duration'

* Update Music Player, Now uses the new functions

* Use resume instead of play, and only only go forward or backwards if sound is playing

* feat: add specular for light data

* fix: update for new GCC

* fix: process exceptions from callbacks

* feat: strip redundant functions/consts

---------

+++ src/ee_tools.c
@@ -3,7 +3,7 @@
 
 void RedirectFunction(void* old, void* new) {
     u32* function = old;
-    u32 patch = new;
+    u32 patch = (u32)new;
 
     FlushCache(2);
 
@@ -17,7 +17,7 @@ void RedirectFunction(void* old, void* new) {
 
 void RedirectCall(void* old, void* new) {
     u32* function = old;
-    u32 patch = new;
+    u32 patch = (u32)new;
 
     FlushCache(2);
